+++ kmertools/src/main.rs
@@ -14,7 +14,8 @@ fn main() {
     match cli.command {
         args::Commands::Comp { command } => match command {
             CompositionCommands::Oligo(command) => {
-                let mut com = OligoComputer::new(command.input, command.output, command.k_size);
+                let mut com =
+                    OligoComputer::new(command.input, command.output, command.k_size as usize);
                 if command.threads > 0 {
                     com.set_threads(command.threads);
                 }
